% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filtrar una matriz para hacer el heatmap con un valor de cuttof establecido}
\usage{
filter_matrix(x, cutoff = 0)
}
\arguments{
\item{x}{A expression matrix with spcies x genes}

\item{cutoff}{A numeric value to subset}
}
\value{
A ComplexHeatmap object filtered
}
\description{
Filtrar una matriz para hacer el heatmap con un valor de cuttof establecido
}
\examples{
expresion_genes <- matrix(rnorm(100), nrow = 10)
rownames(expresion_genes) <- paste0("gene_",letters[1:10])
colnames(expresion_genes) <- paste0("especie_",letters[1:10])
filter_matrix(expresion_genes, cutoff = 0.5)

}
