% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_heatmap.R
\name{subset_heatmap}
\alias{subset_heatmap}
\title{Crea un heatmap con un subgrupo de genes y especies}
\usage{
subset_heatmap(x, especies = NULL, genes = NULL)
}
\arguments{
\item{x}{A expression matrix with especies x genes}

\item{especies}{character() especies to subset}

\item{genes}{character() genes to subset}
}
\value{
A ComplexHeatmap object
}
\description{
Crea un heatmap con un subgrupo de genes y especies
}
\examples{
# create expresion matrix
expresion_genes <- matrix(rnorm(100), nrow = 10)
rownames(expresion_genes) <- paste0("gene_",letters[1:10])
colnames(expresion_genes) <- paste0("especie_",letters[1:10])

subset_heatmap(expresion_genes,
especies = c("especie_a", "especie_b","especie_c"),
genes = c("gene_d","gene_e","gene_f"))

}
