% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Function to filter matrix to an cutoff value}
\usage{
filter_matrix(x, cutoff = 0)
}
\arguments{
\item{x}{A subset of the expression matrix from species x genes}

\item{cutoff}{A default cutoff value set to 0}
}
\value{
A ComplexHeatmap object
}
\description{
Function to filter matrix to an cutoff value
}
\examples{
# create a xpression matrix
expresion_genes <- matrix(rnorm(100), nrow = 10)
rownames(expresion_genes) <- paste0("gene_",letters[1:10])
colnames(expresion_genes) <- paste0("especie_",letters[1:10])

subset_heatmap(expresion_genes,
especies = c("especie_a", "especie_b","especie_c")


}
