% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Crea un heatmap de una matrix de expresion con valores mayores a un cutoff}
\usage{
filter_matrix(x, cutoff = 0)
}
\arguments{
\item{x}{A expression matrix with especies x genes}

\item{cutoff}{numeric()}
}
\value{
A ComplexHeatmap object
}
\description{
Crea un heatmap de una matrix de expresion con valores mayores a un cutoff
}
\examples{
#' # create expresion matrix
expresion_genes <- matrix(rnorm(100), nrow = 10)
rownames(expresion_genes) <- paste0("gene_",letters[1:10])
colnames(expresion_genes) <- paste0("especie_",letters[1:10])

filter_matrix(expresion_genes, 0.5)

}
